Refactor SettingScreen for debug/release variants
Moved SettingScreen implementation to androidDebug and androidRelease source sets, separating debug-only developer options from release builds. Updated androidMain to remove debug-specific UI and logic, ensuring correct behavior per build type.
