+++ core/component/src/androidRelease/kotlin/com/keelim/composeutil/screen/setting/SettingScreen.kt
@@ -1,0 +1,95 @@
+package com.keelim.composeutil.screen.setting
+
+import androidx.compose.foundation.clickable
+import androidx.compose.foundation.layout.Column
+import androidx.compose.foundation.layout.Spacer
+import androidx.compose.foundation.layout.fillMaxWidth
+import androidx.compose.foundation.layout.height
+import androidx.compose.foundation.layout.padding
+import androidx.compose.material3.HorizontalDivider
+import androidx.compose.material3.MaterialTheme
+import androidx.compose.material3.Surface
+import androidx.compose.material3.Text
+import androidx.compose.runtime.Composable
+import androidx.compose.ui.Alignment
+import androidx.compose.ui.Modifier
+import androidx.compose.ui.graphics.Color
+import androidx.compose.ui.unit.dp
+import com.keelim.composeutil.resource.space16
+import com.keelim.composeutil.resource.space4
+import com.keelim.composeutil.resource.space8
+
+@Composable
+fun SettingScreen(
+    modifier: Modifier = Modifier,
+    appName: String = "",
+    clickAction: ((SettingAction) -> Unit)? = null,
+    developerModeDataHolder: Map<String, String>? = null,
+) {
+    Surface(
+        modifier = modifier
+            .padding(
+                horizontal = space8,
+                vertical = space16,
+            ),
+    ) {
+        Column(
+            modifier = Modifier.fillMaxWidth(),
+        ) {
+            Text(
+                text = "$appName 환경설정",
+                style = MaterialTheme.typography.titleSmall,
+            )
+            Spacer(modifier = Modifier.height(20.dp))
+            Text(
+                modifier = Modifier.clickable {
+                    clickAction?.invoke(SettingAction.AlarmSetting)
+                },
+                text = "알림 설정",
+                style = MaterialTheme.typography.titleSmall,
+            )
+            Spacer(modifier = Modifier.height(space4))
+            HorizontalDivider(
+                thickness = 1.dp,
+                color = Color.DarkGray,
+            )
+            Spacer(modifier = Modifier.height(space16))
+            Text(
+                modifier = Modifier.clickable {
+                    clickAction?.invoke(SettingAction.ThemeSetting)
+                },
+                text = "테마 설정",
+                style = MaterialTheme.typography.titleSmall,
+            )
+            Spacer(modifier = Modifier.height(space4))
+            HorizontalDivider(
+                thickness = 1.dp,
+                color = Color.DarkGray,
+            )
+            Spacer(modifier = Modifier.height(space16))
+            Text(
+                modifier = Modifier.clickable {
+                    clickAction?.invoke(SettingAction.Lab)
+                },
+                text = "실험실",
+                style = MaterialTheme.typography.titleSmall,
+            )
+            Spacer(modifier = Modifier.height(space4))
+            HorizontalDivider(
+                thickness = 1.dp,
+                color = Color.DarkGray,
+            )
+            Spacer(modifier = Modifier.height(space8))
+            Text(
+                modifier = Modifier
+                    .align(Alignment.End)
+                    .clickable {
+                        clickAction?.invoke(SettingAction.OtherApp)
+                    },
+                text = "개발자가 만든 다른 앱",
+                style = MaterialTheme.typography.labelSmall,
+                color = Color.DarkGray,
+            )
+        }
+    }
+}
